+++ source/renderer/app/components/news/IncidentOverlay.js
@@ -12,16 +12,11 @@ import styles from './IncidentOverlay.scss';
 type Props = {
   incident: News.News,
   onOpenExternalLink: Function,
+  currentDateFormat: string,
 };
 
 @observer
 export default class IncidentOverlay extends Component<Props> {
-  localizedDateFormat: 'MM/DD/YYYY';
-
-  componentWillMount() {
-    this.localizedDateFormat = moment.localeData().longDateFormat('L');
-  }
-
   contentClickHandler(event: SyntheticMouseEvent<HTMLElement>) {
     const linkUrl = get(event, ['target', 'href']);
     if (linkUrl) {
@@ -46,13 +41,13 @@ export default class IncidentOverlay extends Component<Props> {
   };
 
   render() {
-    const { incident } = this.props;
+    const { incident, currentDateFormat } = this.props;
     const { content, date, action, title } = incident;
     return (
       <div className={styles.component}>
         <h1 className={styles.title}>{title}</h1>
         <span className={styles.date}>
-          {moment(date).format(this.localizedDateFormat)}
+          {moment(date).format(currentDateFormat)}
         </span>
         <div
           className={styles.content}