+++ tests/settings/e2e/steps/select-language.js
@@ -1,12 +1,12 @@
 // @flow
 import { Given, When, Then } from 'cucumber';
 import { expect } from 'chai';
-import { languageSelectionHelpers } from './helpers';
+import { initialSettingsHelpers } from './helpers';
 import type { Daedalus } from '../../../types';
 
 declare var daedalus: Daedalus;
-const { ensureLanguageIsSelected } = languageSelectionHelpers;
-const LANGUAGE_SELECTION_FORM = '.LanguageSelectionForm_component';
+const { ensureLanguageIsSelected } = initialSettingsHelpers;
+const INITIAL_SETTINGS_FORM = '.InitialSettings_component';
 
 Given(/^I have selected English language$/, async function() {
   await ensureLanguageIsSelected(this.client, {
@@ -21,12 +21,12 @@ Given(/^I dont have a language set$/, async function() {
 });
 
 When(/^I am on the language selection screen$/, function() {
-  return this.client.waitForVisible('.LanguageSelectionForm_component');
+  return this.client.waitForVisible('.InitialSettings_component');
 });
 
 When(/^I open language selection dropdown$/, function() {
   return this.waitAndClick(
-    '.LanguageSelectionForm_component .SimpleInput_input'
+    '.InitialSettings_component .SimpleInput_input'
   );
 });
 
@@ -37,11 +37,11 @@ When(/^I select Japanese language$/, function() {
 });
 
 When(/^I submit the language selection form$/, function() {
-  return this.waitAndClick('.LanguageSelectionForm_submitButton');
+  return this.waitAndClick('.ProfileSettingsForm_submitButton');
 });
 
 Then(/^I should not see the language selection screen anymore$/, function() {
-  return this.client.waitForVisible(LANGUAGE_SELECTION_FORM, null, true);
+  return this.client.waitForVisible(INITIAL_SETTINGS_FORM, null, true);
 });
 
 Then(/^I should have Japanese language set$/, async function() {